Adjusted grid layout on countries page.
