+++ app/post/[slug]/page.tsx
@@ -25,8 +25,8 @@ export default async function Post({ params }: PostPageParams) {
 
   return (
     <article>
-      <Section spacing="none" className="relative">
-        <div className="flex flex-col col-span-8 gap-8 px-5 py-8 lg:relative lg:col-start-3 lg:px-0 md:py-16 lg:py-24">
+      <Section spacing="none">
+        <div className="flex flex-col col-span-8 gap-8 px-5 py-8 lg:col-start-3 lg:px-0 md:py-16 lg:py-24">
           {post.countries && post.countries.length > 0 && (
             <CountryList countries={post.countries} large />
           )}
