+++ src/components/Intro.tsx
@@ -5,7 +5,7 @@ import Link from 'next/link';
 import { Section } from 'components';
 
 const Intro: FC = () => (
-  <Section spacing="sm">
+  <Section spacing="none">
     <div className="flex flex-col justify-center col-span-12 px-5 mb-16 lg:col-start-3 md:min-h-most lg:px-0 md:mb-0">
       <p className="text-step-5">
         <span className="inline-block animate-wave">👋</span>
